refactor: move faq logic to hook
